+++ src/lib/components/RedSeaFeatures.svelte
